Move Player from types-todo to better player-wip
